Change(package): "Move class in package clearly"
- Organize classes in a jumbled package.
- FluctuatedDialogFragment.class name is changed into ClassChangedDialogFragment.

Resolves : #61
