+++ app/src/main/java/com/dnd/sixth/lmsservice/di/modules/ViewModelModules.kt
@@ -5,12 +5,12 @@ import com.dnd.sixth.lmsservice.presentation.feedback.WriteFeedBackViewModel
 import com.dnd.sixth.lmsservice.presentation.lesson.progress.ClassProgressViewModel
 import com.dnd.sixth.lmsservice.presentation.login.LoginViewModel
 import com.dnd.sixth.lmsservice.presentation.main.MainViewModel
-import com.dnd.sixth.lmsservice.presentation.main.home.ClassManageViewModel
-import com.dnd.sixth.lmsservice.presentation.main.home.making.MakeClassViewModel
-import com.dnd.sixth.lmsservice.presentation.main.schedule.add.ScheduleAddViewModel
-import com.dnd.sixth.lmsservice.presentation.main.schedule.calendar.HomeCalendarViewModel
-import com.dnd.sixth.lmsservice.presentation.main.schedule.calendar.HomeClassViewModel
-import com.dnd.sixth.lmsservice.presentation.main.schedule.calendar.HomeViewModel
+import com.dnd.sixth.lmsservice.presentation.home.classes.ClassManageViewModel
+import com.dnd.sixth.lmsservice.presentation.home.classes.manage.making.MakeClassViewModel
+import com.dnd.sixth.lmsservice.presentation.home.classes.calendar.add.ScheduleAddViewModel
+import com.dnd.sixth.lmsservice.presentation.home.classes.calendar.HomeCalendarViewModel
+import com.dnd.sixth.lmsservice.presentation.home.classes.manage.HomeClassViewModel
+import com.dnd.sixth.lmsservice.presentation.home.HomeViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
