+++ app/src/main/java/com/dnd/sixth/lmsservice/presentation/adapter/viewpager/ClassManageViewPagerAdapter.kt
@@ -1,14 +1,15 @@
 package com.dnd.sixth.lmsservice.presentation.adapter.viewpager
 
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.dnd.sixth.lmsservice.presentation.main.schedule.calendar.ClassCalendarFragment
-import com.dnd.sixth.lmsservice.presentation.main.schedule.calendar.ClassManageActivity
+import com.dnd.sixth.lmsservice.presentation.home.classes.calendar.HomeCalendarFragment
+import com.dnd.sixth.lmsservice.presentation.home.classes.manage.HomeClassFragment
 
-class ClassManageViewPagerAdapter(activity: ClassManageActivity) : FragmentStateAdapter(activity) {
+class ClassManageViewPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activity) {
 
     // ViewPager2에 연결할 Fragment 생성
-    private val fragmentList = listOf<Fragment>(ClassCalendarFragment(), ClassCalendarFragment(), ClassCalendarFragment())
+    private val fragmentList = listOf<Fragment>(HomeClassFragment(), HomeCalendarFragment())
 
     // ViesPager2에서 노출시킬 Fragment 개수 설정
     override fun getItemCount(): Int {
