+++ app/src/main/java/com/dnd/sixth/lmsservice/presentation/home/classes/ClassManageViewModel.kt
@@ -1,4 +1,4 @@
-package com.dnd.sixth.lmsservice.presentation.main.home
+package com.dnd.sixth.lmsservice.presentation.home.classes
 
 import com.dnd.sixth.lmsservice.presentation.base.BaseViewModel
