+++ app/src/main/java/com/dnd/sixth/lmsservice/presentation/home/classes/manage/making/MakeClassActivity.kt
@@ -1,4 +1,4 @@
-package com.dnd.sixth.lmsservice.presentation.main.home.making
+package com.dnd.sixth.lmsservice.presentation.home.classes.manage.making
 
 import android.view.MenuItem
 import com.dnd.sixth.lmsservice.R
