+++ app/src/main/res/layout/activity_make_class.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.dnd.sixth.lmsservice.presentation.main.home.making.MakeClassViewModel" />
+            type="com.dnd.sixth.lmsservice.presentation.home.classes.manage.making.MakeClassViewModel" />
     </data>
 
 
@@ -19,7 +19,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            tools:context=".presentation.main.home.making.MakeClassActivity">
+            tools:context=".presentation.home.classes.manage.making.MakeClassActivity">
 
             <com.google.android.material.appbar.MaterialToolbar
                 android:id="@+id/toolbar"
