+++ app/src/main/res/layout/dialog_push_time_picker.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.dnd.sixth.lmsservice.presentation.main.schedule.notipicker.PushTimePickerViewModel" />
+            type="com.dnd.sixth.lmsservice.presentation.home.classes.calendar.notipicker.PushTimePickerViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
