+++ app/src/main/res/layout/fragment_schedule_edit_dialog.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.dnd.sixth.lmsservice.presentation.main.schedule.edit.ScheduleEditViewModel" />
+            type="com.dnd.sixth.lmsservice.presentation.home.classes.calendar.edit.ScheduleEditViewModel" />
     </data>
 
     <ScrollView
@@ -21,7 +21,7 @@
             android:layout_height="match_parent"
             android:paddingHorizontal="16dp"
             android:paddingBottom="16dp"
-            tools:context=".presentation.main.schedule.add.ScheduleAddActivity">
+            tools:context=".presentation.home.classes.calendar.add.ScheduleAddActivity">
 
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/toolbar"
